refactor: move message-pickup directory (#1650)
