+++ packages/core/src/agent/BaseAgent.ts
@@ -3,7 +3,7 @@ import type { AgentApi, CustomOrDefaultApi, EmptyModuleMap, ModulesMap, WithoutD
 import type { TransportSession } from './TransportService'
 import type { Logger } from '../logger'
 import type { CredentialsModule } from '../modules/credentials'
-import type { MessagePickupModule } from '../modules/message-pìckup'
+import type { MessagePickupModule } from '../modules/message-pickup'
 import type { ProofsModule } from '../modules/proofs'
 import type { DependencyManager } from '../plugins'
 
@@ -14,7 +14,7 @@ import { CredentialsApi } from '../modules/credentials'
 import { DidsApi } from '../modules/dids'
 import { DiscoverFeaturesApi } from '../modules/discover-features'
 import { GenericRecordsApi } from '../modules/generic-records'
-import { MessagePickupApi } from '../modules/message-pìckup/MessagePickupApi'
+import { MessagePickupApi } from '../modules/message-pickup/MessagePickupApi'
 import { OutOfBandApi } from '../modules/oob'
 import { ProofsApi } from '../modules/proofs'
 import { MediatorApi, MediationRecipientApi } from '../modules/routing'
