+++ packages/core/src/agent/__tests__/AgentModules.test.ts
@@ -7,7 +7,7 @@ import { CredentialsModule } from '../../modules/credentials'
 import { DidsModule } from '../../modules/dids'
 import { DiscoverFeaturesModule } from '../../modules/discover-features'
 import { GenericRecordsModule } from '../../modules/generic-records'
-import { MessagePickupModule } from '../../modules/message-pìckup'
+import { MessagePickupModule } from '../../modules/message-pickup'
 import { OutOfBandModule } from '../../modules/oob'
 import { ProofsModule } from '../../modules/proofs'
 import { MediationRecipientModule, MediatorModule } from '../../modules/routing'
