+++ packages/core/src/modules/message-pickup/MessagePickupModuleConfig.ts
