+++ packages/core/src/modules/message-pickup/__tests__/pickup.test.ts
