+++ packages/core/src/modules/message-pickup/protocol/BaseMessagePickupProtocol.ts
