+++ packages/core/src/modules/message-pickup/protocol/MessagePickupProtocolOptions.ts
