+++ packages/core/src/modules/message-pickup/protocol/v1/V1MessagePickupProtocol.ts
