+++ packages/core/src/modules/message-pickup/protocol/v1/messages/V1BatchMessage.ts
