+++ packages/core/src/modules/message-pickup/protocol/v2/V2MessagePickupProtocol.ts
