+++ packages/core/src/modules/message-pickup/protocol/v2/index.ts
