+++ packages/core/src/modules/message-pickup/protocol/v2/messages/V2MessageDeliveryMessage.ts
